added new policy files

+++ alg/feature_func.py
@@ -2,7 +2,7 @@ import numpy as np
 
 DESIRED_GOAL = np.array([ 0.  , -0.2 ,  0.02])
 
-def get_mean(states,index_min,index_max):
+def get_mean(states, index_min, index_max):
     '''
     Returns the mean of a part of an observation over all states in a trajectory
     '''
@@ -14,10 +14,11 @@ def get_euclidean_distance(states, desired_goal):
     Returns the mean distance over all states in a trajectory
     '''
     achieved_goal = np.array([state[-3:] for state in states])
+    
     distances = np.linalg.norm(achieved_goal - desired_goal, axis=1)
     return distances.mean()
 
-def get_min(states,index):
+def get_min(states, index):
     '''
     Returns the minimum gripper width over all states
     '''
@@ -29,13 +30,11 @@ def feature_function(traj):
     This function gives the following features of a trajectory: distance, 
     velocity, angle, rotation and finger width.
     '''
-
     states = np.array([state_action_pair[0] for state_action_pair in traj])
-
-    hand_position = get_mean(states,0,3)
+    hand_position = get_mean(states, 0, 3)
     banana_distance = get_euclidean_distance(states, DESIRED_GOAL)
-    banana_rotation = get_mean(states,10,13)
-    banana_velocity = get_mean(states,13,16)
-    min_finger_width = get_min(states,6)
+    banana_rotation = get_mean(states, 10, 13)
+    banana_velocity = get_mean(states, 13, 16)
+    min_finger_width = get_min(states, 6)
 
-    return np.array([hand_position,banana_distance,banana_rotation,banana_velocity,min_finger_width])
+    return np.array([hand_position, banana_distance, banana_rotation, banana_velocity, min_finger_width], dtype=np.float32)
